services/user_service/requirements.txt
fastapi==0.114.0        # Фреймворк для микросервиса
uvicorn==0.31.0        # ASGI сервер для запуска FastAPI
pydantic==2.9.2       # Для валидации данных и создания моделей
sqlalchemy==2.0.35      # Для работы с базой данных (ORM)
alembic==1.13.3        # Миграции базы данных
passlib==1.7.4         # Для хэширования паролей
# bcrypt==4.0.1           # Возможно потребуется для реализации JWT+хэширования
python-jose==3.3.0     # Для работы с JWT токенами
asyncpg==0.29.0        # Асинхронный движок для sql
databases[postgresql]==0.9.0  # Для работы с СУБД postgresql
python-dotenv==1.0.1    # Работа с env файлами, для защиты данных проекта
asyncio==3.4.3         # Работа асинхронными методами именно 3.4.3, т.к. ниже версии не подходят